requirements.txt
black==23.3.0
click==8.1.3
lxml==4.9.2
mypy-extensions==1.0.0
numpy==1.24.2
packaging==23.1
pathspec==0.11.1
platformdirs==3.2.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.0
pyqt5-tools==5.15.9.3.3
python-docx==0.8.11
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
